move log out of internal pkg
