+++ tools/ghx/log/default.go
