+++ tools/ghx/log/log.go
