+++ tools/ghx/runner/runner.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 
 	"github.com/aweris/gale/internal/core"
-	"github.com/aweris/gale/tools/ghx/internal/log"
+	"github.com/aweris/gale/tools/ghx/log"
 )
 
 type Runner struct {
