+++ tools/ghx/runner/runner_step.go
@@ -9,7 +9,7 @@ import (
 	"github.com/aweris/gale/internal/core"
 	"github.com/aweris/gale/internal/fs"
 	"github.com/aweris/gale/tools/ghx/actions"
-	"github.com/aweris/gale/tools/ghx/internal/log"
+	"github.com/aweris/gale/tools/ghx/log"
 )
 
 type Step interface {
